Partition sign_psbt.c into smaller submodules
Also, updated the copyright notices, or added where missing.
