+++ src/handler/sign_psbt/transaction_display.h
@@ -17,22 +17,30 @@
 
 #pragma once
 
-/* Local headers */
+#include <stdbool.h>
+#include <stdint.h>
+
+#include "bitvector.h"
+#include "constants.h"
 #include "dispatcher.h"
-#include "merkle.h"
-#include "sign_psbt_cache.h"
-#include "wallet.h"
+#include "sign_psbt.h"
+
+// UI text labels defined in src/ui (used to set the "loading"/"signing"
+// processing screen).
+extern const char GA_LOADING_TRANSACTION[];
+extern const char GA_SIGNING_TRANSACTION[];
 
 /**
- * TODO
+ * Drives the user-facing transaction confirmation flow:
+ * - displays warnings (external inputs, non-default sighash, missing
+ *   non-witness UTXO);
+ * - displays each external output and the fees;
+ * - asks for final user approval.
+ *
+ * Returns true if the user approved, false otherwise (in which case an error
+ * status word has already been sent).
  */
-int compare_wallet_script_at_path(dispatcher_context_t *dispatcher_context,
-                                  sign_psbt_cache_t *sign_psbt_cache,
-                                  uint32_t change,
-                                  uint32_t address_index,
-                                  const policy_node_t *policy,
-                                  int wallet_version,
-                                  const uint8_t keys_merkle_root[static 32],
-                                  uint32_t n_keys,
-                                  const uint8_t expected_script[],
-                                  size_t expected_script_len);
+bool display_transaction(
+    dispatcher_context_t *dc,
+    sign_psbt_state_t *st,
+    const uint8_t internal_outputs[static BITVECTOR_REAL_SIZE(MAX_N_OUTPUTS_CAN_SIGN)]);
